+++ app/src/main/java/com/orange/ods/app/ui/modules/ModulesNavGraph.kt
@@ -13,20 +13,28 @@
 package com.orange.ods.app.ui.modules
 
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.orange.ods.app.ui.CustomAppBarConfiguration
+import com.orange.ods.app.ui.LocalAppBarManager
 import com.orange.ods.app.ui.modules.ModulesNavigation.AboutCustomizationRoute
 import com.orange.ods.app.ui.modules.ModulesNavigation.EmptyStateCustomizationRoute
 import com.orange.ods.app.ui.modules.ModulesNavigation.EmptyStateDemoRoute
+import com.orange.ods.app.ui.modules.ModulesNavigation.MoreAppsCustomizationRoute
 import com.orange.ods.app.ui.modules.about.AboutCustomizationScreen
 import com.orange.ods.app.ui.modules.about.AboutCustomizationViewModel
 import com.orange.ods.app.ui.modules.emptystate.EmptyStateCustomizationScreen
 import com.orange.ods.app.ui.modules.emptystate.EmptyStateCustomizationViewModel
 import com.orange.ods.app.ui.modules.emptystate.EmptyStateDemoScreen
+import com.orange.ods.app.ui.modules.moreapps.MoreAppsCustomizationScreen
+import com.orange.ods.app.ui.modules.moreapps.MoreAppsCustomizationViewModel
 import com.orange.ods.app.ui.navigateToElement
+import com.orange.ods.module.moreapps.ui.configuration.OdsMoreAppsConfiguration
+import com.orange.ods.module.moreapps.ui.navigation.odsMoreAppsGraph
 
 /**
  * Modules demo destinations.
@@ -34,11 +42,18 @@ import com.orange.ods.app.ui.navigateToElement
 object ModulesNavigation {
     const val AboutCustomizationRoute = "module/about/customization"
 
+    const val MoreAppsCustomizationRoute = "module/moreApps/customization"
+
     const val EmptyStateCustomizationRoute = "module/emptyState/customization"
     const val EmptyStateDemoRoute = "module/emptyState/demo"
 }
 
-fun NavGraphBuilder.addModulesGraph(navController: NavController, navigateToAboutDemo: () -> Unit) {
+fun NavGraphBuilder.addModulesGraph(
+    navController: NavController,
+    navigateToAboutDemo: () -> Unit,
+    navigateToMoreAppsDemo: () -> Unit,
+    moreAppsConfiguration: () -> OdsMoreAppsConfiguration
+) {
     composable(route = AboutCustomizationRoute) { _ ->
         val viewModelStoreOwner = LocalContext.current as ViewModelStoreOwner
         val viewModel = viewModel<AboutCustomizationViewModel>(viewModelStoreOwner)
@@ -55,4 +70,20 @@ fun NavGraphBuilder.addModulesGraph(navController: NavController, navigateToAbou
             EmptyStateDemoScreen(viewModel = viewModel)
         }
     }
+
+    composable(route = MoreAppsCustomizationRoute) { _ ->
+        val appBarManager = LocalAppBarManager.current
+        val appBarTitle = stringResource(id = com.orange.ods.module.moreapps.R.string.odsMoreApps_title)
+        val viewModelStoreOwner = LocalContext.current as ViewModelStoreOwner
+        val viewModel = viewModel<MoreAppsCustomizationViewModel>(viewModelStoreOwner)
+        MoreAppsCustomizationScreen(
+            navigateToMoreAppsDemo = {
+                navigateToMoreAppsDemo()
+                appBarManager.setCustomAppBar(CustomAppBarConfiguration(title = appBarTitle, actionCount = 0))
+            },
+            viewModel = viewModel
+        )
+    }
+
+    odsMoreAppsGraph(moreAppsConfiguration)
 }