+++ module-more-apps/src/main/java/com/orange/ods/module/moreapps/domain/OdsMoreAppsDomainModule.kt
@@ -1,0 +1,28 @@
+/*
+ * Software Name: Orange Design System
+ * SPDX-FileCopyrightText: Copyright (c) Orange SA
+ * SPDX-License-Identifier: MIT
+ *
+ * This software is distributed under the MIT license,
+ * the text of which is available at https://opensource.org/license/MIT/
+ * or see the "LICENSE" file for more details.
+ *
+ * Software description: Android library of reusable graphical components
+ */
+
+package com.orange.ods.module.moreapps.domain
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal object OdsMoreAppsDomainModule {
+
+    @Singleton
+    @Provides
+    fun provideMoreAppsService(moreAppsRepository: MoreAppsRepository): MoreAppsService = MoreAppsServiceImpl(moreAppsRepository)
+}